Time: 3 ms (51.28%) | Memory: 10.1 MB (38.94%) - LeetSync

+++ 2076-sum-of-digits-of-string-after-convert/sum-of-digits-of-string-after-convert.cpp
@@ -16,7 +16,7 @@ class Solution
             for (int transform = 0; transform < k; transform++) {
                 sum = 0;
                 for(char c:alph){
-                    sum += (int)c - 48;
+                    sum += c - 48;
                 }
 
                 alph = to_string(sum);
